backend/audio2face/requirements.txt
sounddevice
python-dotenv 
pydub
websockets==12.0
grpcio==1.51.3
numpy==1.22.4
soundfile==0.12.1
protobuf==3.20.0
mediapipe==0.10.11
flask==2.3.3
requests==2.31.0